backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
django-debug-toolbar==4.4.6
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2>=2.9,<3.0
